feat: old db delation management

+++ clearOldDB.sh
@@ -1,3 +1,11 @@
-DIR="./bkp/$DB_NAME/"
-KEEP=72
-find "$DIR" -type f -printf "%T@ %p\n" | sort -n | head -n -"$KEEP" | cut -d' ' -f2- | xargs -d '\n' rm -f
+for file in $PWD/envs/{*,.*}
+do
+    if [ -f "$file" ]; then
+      source $file
+      DIR="./bkp/$DB_NAME/"
+      KEEP=72
+      find "$DIR" -type f -printf "%T@ %p\n" | sort -n | head -n -"$KEEP" | cut -d' ' -f2- | xargs -d '\n' rm -f
+    fi
+done
+
+
